fifo_buffer changed from int to uint8_t

+++ src/main.cpp
@@ -166,7 +166,7 @@ MPU6050 mpu;
 // MPU-6050 > DMP
 bool dmp_ready = false; // on board computer to do math for orientation
 int packet_size;
-int fifo_buffer[64];
+uint8_t fifo_buffer[64];
 
 // MPU-6050 > Orientation
 Quaternion quaternion;
